Move MCPDeprecationWarning to mcp.shared.exceptions per @Kludex review
Drops the dedicated _warnings.py — the side-effect-free-module constraint that
justified it is moot now that the filter matches by message only.
